requirements.txt
Django==5.1.6
djangorestframework==3.14.0
django-cors-headers==3.13.0
python-decouple==3.8
psycopg2-binary==2.9.6
